requirements.txt
bcrypt==3.2.0
cffi==1.14.6
click==7.1.2
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.3
Flask==1.1.4
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.6
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Markdown==0.3
Flask-SQLAlchemy==2.5.1
Flask-Toastr==0.5.4
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
importlib-metadata==4.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==2.0.1
Pillow==8.3.1
pip==21.2.4
pycparser==2.20
pyOpenSSL==20.0.1
python-slugify==5.0.2
setuptools==47.1.0
six==1.16.0
SQLAlchemy==1.4.22
text-unidecode==1.3
typing-extensions==3.10.0.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.5.0
